+++ changelog.md
@@ -6,6 +6,7 @@
 * `[process]` Fixed threads marking
 * `[fmtutil]` Added method `CountDigits` for counting the number of digits in integer
 * `[terminal]` Now `PrintWarnMessage` and `PrintErrorMessage` prints messages to stderr
+* `[usage]` Added optional command arguments support
 
 #### v5.6.0